requirements.txt
pyodbc
networkx
matplotlib
pyyaml

# Optional dependencies
# For testing
pytest
# For database test containers
testcontainers
